Added mouse tracking to world state

+++ ProjM1/src/GameState.h
@@ -17,8 +17,8 @@ public:
     float worldTime = 0; //Current time in the world total
     float deltaTime = 0; //Time since last update
     int waveNumber = 0;  //How many waves you've survived
-    float mouseX = 0;    //Current position of mouse (X)
-    float mouseY = 0;    //Current position of mouse (Y)
+    double mouseX = 0;    //Current position of mouse (X)
+    double mouseY = 0;    //Current position of mouse (Y)
     GLFWwindow *window = NULL;
     //use glfwGetKey (GLFWwindow *window, int key) to check key input
 };
